+++ layers/GraphQLByPoP/packages/graphql-server/src/TypeResolvers/ObjectType/AbstractSchemaElementExtensionsObjectTypeResolver.php
@@ -1,0 +1,51 @@
+<?php
+
+declare(strict_types=1);
+
+namespace GraphQLByPoP\GraphQLServer\TypeResolvers\ObjectType;
+
+use GraphQLByPoP\GraphQLServer\ObjectModels\SchemaDefinitionReferenceObjectInterface;
+use GraphQLByPoP\GraphQLServer\RelationalTypeDataLoaders\ObjectType\SchemaDefinitionReferenceTypeDataLoader;
+use PoP\ComponentModel\RelationalTypeDataLoaders\RelationalTypeDataLoaderInterface;
+
+abstract class AbstractSchemaElementExtensionsObjectTypeResolver extends AbstractIntrospectionObjectTypeResolver
+{
+    private ?SchemaDefinitionReferenceTypeDataLoader $schemaDefinitionReferenceTypeDataLoader = null;
+
+    final public function setSchemaDefinitionReferenceTypeDataLoader(SchemaDefinitionReferenceTypeDataLoader $schemaDefinitionReferenceTypeDataLoader): void
+    {
+        $this->schemaDefinitionReferenceTypeDataLoader = $schemaDefinitionReferenceTypeDataLoader;
+    }
+    final protected function getSchemaDefinitionReferenceTypeDataLoader(): SchemaDefinitionReferenceTypeDataLoader
+    {
+        return $this->schemaDefinitionReferenceTypeDataLoader ??= $this->instanceManager->getInstance(SchemaDefinitionReferenceTypeDataLoader::class);
+    }
+
+    /**
+     * Introspection names must start with "__".
+     * However, when doing so, graphql-js throws an error:
+     *
+     * @see https://github.com/graphql-java/graphql-java/pull/2221#issuecomment-808044041
+     *
+     * To avoid it, prepend it with "_", as a temporary solution, until
+     * the GraphQL spec and graphql-js deal with the issue.
+     *
+     * @see https://github.com/graphql/graphql-spec/issues/300#issuecomment-808047303
+     */
+    final public function getTypeName(): string
+    {
+        return '_' . $this->getIntrospectionTypeName();
+    }
+    abstract protected function getIntrospectionTypeName(): string;
+
+    public function getID(object $object): string | int | null
+    {
+        /** @var SchemaDefinitionReferenceObjectInterface */
+        return $object->getID();
+    }
+
+    public function getRelationalTypeDataLoader(): RelationalTypeDataLoaderInterface
+    {
+        return $this->getSchemaDefinitionReferenceTypeDataLoader();
+    }
+}
